requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cloudinary==1.29.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
Pillow==9.2.0
PyJWT==1.7.1
Pillow==9.2.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9